Fix #8 Move CI to GitHub actions
Also extends the linting to the application of yamllint.

Also adds a CONTRIBUTING.md, to explain the approach to testing and linting.

Also applies HLint's suggestions.

Also applies yamllint's suggestions.

Also adds an empty `.hlint.yaml` file, in case it is needed in the future.

Deletes the Azure CI files.